sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orchestra-ai"
version = "0.1.0"
description = "Python SDK for OrchestraAI - Observability & Control Plane for AI Agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "OrchestraAI Team", email = "team@orchestra-ai.dev" }
]
keywords = [
    "ai",
    "agents",
    "observability",
    "tracing",
    "llm",
    "langgraph",
    "openai",
    "crewai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
langgraph = [
    "langgraph>=0.0.50",
]
openai = [
    "openai>=1.0.0",
]
crewai = [
    "crewai>=0.30.0",
]
all = [
    "langgraph>=0.0.50",
    "openai>=1.0.0",
    "crewai>=0.30.0",
]

[project.urls]
Homepage = "https://github.com/your-org/orchestra-ai"
Documentation = "https://docs.orchestra-ai.dev"
Repository = "https://github.com/your-org/orchestra-ai"

[tool.hatch.build.targets.wheel]
packages = ["orchestra_ai"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
